+++ src/main/java/com/dnsimple/data/ZoneRecordBatchDelete.java
@@ -1,0 +1,13 @@
+package com.dnsimple.data;
+
+public class ZoneRecordBatchDelete {
+    private final Long id;
+
+    public ZoneRecordBatchDelete(Long id) {
+        this.id = id;
+    }
+
+    public Long getId() {
+        return id;
+    }
+}